some fixes

+++ libft/ft_libc/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: flcarval <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 12:35:41 by flcarval          #+#    #+#             */
-/*   Updated: 2021/11/22 12:35:42 by flcarval         ###   ########.fr       */
+/*   Updated: 2021/11/23 14:50:59 by flcarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
-	int 	t;
+	size_t	i;
+	int	t;
 
 	i = 0;
 	while (i < n && (s1[i] || s2[i]))
